move supertest _handle subprocess test into node-http.test.ts
