+++ test/js/node/http/node-http.test.ts
@@ -4053,6 +4053,36 @@ it("http.Server exposes _handle like net.Server (null/truthy/null over listen/cl
   }
 });
 
+it("process exits after supertest-style server._handle gated close()", async () => {
+  // https://github.com/oven-sh/bun/issues/13087, https://github.com/oven-sh/bun/issues/23648
+  const fixture = /* js */ `
+    const http = require("node:http");
+    const server = http.createServer((req, res) => res.end("hi"));
+    server.listen(0, () => {
+      const port = server.address().port;
+      http.get("http://127.0.0.1:" + port + "/", (res) => {
+        res.resume();
+        res.on("end", () => {
+          // supertest's end(): if (server && server._handle) return server.close(cb); cb();
+          const done = () => console.log("done", server._handle == null ? "ok" : "leaked");
+          if (server && server._handle) server.close(done);
+          else done();
+        });
+      });
+    });
+  `;
+  await using proc = Bun.spawn({
+    cmd: [bunExe(), "-e", fixture],
+    env: bunEnv,
+    stdout: "pipe",
+    stderr: "pipe",
+  });
+  const [stdout, stderr, exitCode] = await Promise.all([proc.stdout.text(), proc.stderr.text(), proc.exited]);
+  expect(stderr).toBe("");
+  expect(stdout.trim()).toBe("done ok");
+  expect(exitCode).toBe(0);
+});
+
 it("OutgoingMessage outputData is per-instance and _flushOutput is defined", () => {
   expect(typeof OutgoingMessage.prototype._flushOutput).toBe("function");
